Moving third party canon folder out or package folder
