+++ canon/CameraModel.py
