+++ canon/EDSDKLib.py
